requirements.txt
google-generativeai>=0.3.0  # For Gemini API
google-cloud-texttospeech>=2.14.1  # For Google TTS
moviepy>=1.0.3  # For video editing
python-dotenv>=1.0.0  # For environment variable management
Pillow>=9.5.0  # For image processing
requests>=2.31.0  # For HTTP requests
pydantic>=2.4.0  # For data validation